feat(FederatedStatistics.java): add stats (count and time) for federated lookup table entries and get requests 	add different counts for the different data types in put requests 	add a bytes count for put requests of matrices and frames
refactor(Statistics.java): divide related stats and their methods into own static classes

feat(utils/stats/**.java): create static classes for related statistics

refactor(**.java): change all calls to statistics according to the outsourced stats into new static classes
