+++ src/main/java/org/apache/sysds/hops/codegen/SpoofCompiler.java
@@ -96,7 +96,7 @@ import org.apache.sysds.runtime.lineage.LineageItemUtils;
 import org.apache.sysds.runtime.matrix.data.Pair;
 import org.apache.sysds.utils.Explain;
 import org.apache.sysds.utils.NativeHelper;
-import org.apache.sysds.utils.Statistics;
+import org.apache.sysds.utils.stats.CodegenStatistics;
 
 import java.io.File;
 import java.io.IOException;
@@ -563,7 +563,7 @@ public class SpoofCompiler {
 				}
 				else {
 					if( DMLScript.STATISTICS ) 
-						Statistics.incrementCodegenOpCacheHits();
+						CodegenStatistics.incrementCodegenOpCacheHits();
 					if(CodegenUtils.getCUDAopID(cla.getName()) != null) {
 						tmp.getValue().setGeneratorAPI(GeneratorAPI.CUDA);
 						tmp.getValue().setVarName(cla.getName().split("\\.")[1]);
@@ -586,7 +586,7 @@ public class SpoofCompiler {
 					clas.put(cplan.getKey(), new Pair<Hop[], Class<?>>(tmp.getKey(), cla));
 				}
 				if( DMLScript.STATISTICS )
-					Statistics.incrementCodegenOpCacheTotal();
+					CodegenStatistics.incrementCodegenOpCacheTotal();
 			}
 			
 			//create modified hop dag (operator replacement and CSE)
@@ -612,8 +612,8 @@ public class SpoofCompiler {
 		}
 		
 		if( DMLScript.STATISTICS ) {
-			Statistics.incrementCodegenDAGCompile();
-			Statistics.incrementCodegenCompileTime(System.nanoTime()-t0);
+			CodegenStatistics.incrementCodegenDAGCompile();
+			CodegenStatistics.incrementCodegenCompileTime(System.nanoTime()-t0);
 		}
 		
 		Hop.resetVisitStatus(roots);
@@ -738,7 +738,7 @@ public class SpoofCompiler {
 			if( tmp != null ) {
 				cplans.put(hop.getHopID(), tmp);
 				if (DMLScript.STATISTICS)
-					Statistics.incrementCodegenCPlanCompile(1);
+					CodegenStatistics.incrementCodegenCPlanCompile(1);
 			}
 		}
