+++ src/main/java/org/apache/sysds/runtime/codegen/CodegenUtils.java
@@ -31,7 +31,7 @@ import org.apache.sysds.runtime.codegen.SpoofOperator.SideInputSparseCell;
 import org.apache.sysds.runtime.io.IOUtilFunctions;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
 import org.apache.sysds.runtime.util.LocalFileUtils;
-import org.apache.sysds.utils.Statistics;
+import org.apache.sysds.utils.stats.CodegenStatistics;
 import org.codehaus.janino.SimpleCompiler;
 
 import javax.tools.Diagnostic;
@@ -88,8 +88,8 @@ public class CodegenUtils
 		_cache.put(name, ret);
 
 		if( DMLScript.STATISTICS ) {
-			Statistics.incrementCodegenClassCompile();
-			Statistics.incrementCodegenClassCompileTime(System.nanoTime()-t0);
+			CodegenStatistics.incrementCodegenClassCompile();
+			CodegenStatistics.incrementCodegenClassCompileTime(System.nanoTime()-t0);
 		}
 
 		return ret;
