+++ src/main/java/org/apache/sysds/runtime/transform/encode/ColumnEncoderRecode.java
@@ -37,7 +37,7 @@ import org.apache.sysds.lops.Lop;
 import org.apache.sysds.runtime.controlprogram.caching.CacheBlock;
 import org.apache.sysds.runtime.matrix.data.FrameBlock;
 import org.apache.sysds.runtime.matrix.data.MatrixBlock;
-import org.apache.sysds.utils.Statistics;
+import org.apache.sysds.utils.stats.TransformStatistics;
 
 public class ColumnEncoderRecode extends ColumnEncoder {
 	private static final long serialVersionUID = 8213163881283341874L;
@@ -140,7 +140,7 @@ public class ColumnEncoderRecode extends ColumnEncoder {
 		long t0 = DMLScript.STATISTICS ? System.nanoTime() : 0;
 		makeRcdMap(in, _rcdMap, _colID, 0, in.getNumRows());
 		if(DMLScript.STATISTICS){
-			Statistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
+			TransformStatistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
 		}
 	}
 
@@ -347,7 +347,7 @@ public class ColumnEncoderRecode extends ColumnEncoder {
 				return null;
 			_encoder.applySparse(_input, _out, _outputCol, _startRow, _blk);
 			if(DMLScript.STATISTICS){
-				Statistics.incTransformRecodeApplyTime(System.nanoTime() - t0);
+				TransformStatistics.incTransformRecodeApplyTime(System.nanoTime() - t0);
 			}
 			return null;
 		}
@@ -388,7 +388,7 @@ public class ColumnEncoderRecode extends ColumnEncoder {
 				_partialMaps.put(_startRow, partialMap);
 			}
 			if(DMLScript.STATISTICS){
-				Statistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
+				TransformStatistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
 			}
 			return null;
 		}
@@ -421,7 +421,7 @@ public class ColumnEncoderRecode extends ColumnEncoder {
 			});
 			_encoder._rcdMap = rcdMap;
 			if(DMLScript.STATISTICS){
-				Statistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
+				TransformStatistics.incTransformRecodeBuildTime(System.nanoTime() - t0);
 			}
 			return null;
 		}
