+++ src/main/java/org/apache/sysds/utils/Statistics.java
@@ -25,10 +25,7 @@ import org.apache.sysds.api.DMLScript;
 import org.apache.sysds.conf.ConfigurationManager;
 import org.apache.sysds.hops.OptimizerUtils;
 import org.apache.sysds.runtime.controlprogram.caching.CacheStatistics;
-import org.apache.sysds.runtime.controlprogram.context.SparkExecutionContext;
-import org.apache.sysds.runtime.controlprogram.federated.FederatedRequest.RequestType;
 import org.apache.sysds.runtime.controlprogram.federated.FederatedStatistics;
-import org.apache.sysds.runtime.controlprogram.parfor.stat.Timing;
 import org.apache.sysds.runtime.instructions.Instruction;
 import org.apache.sysds.runtime.instructions.InstructionUtils;
 import org.apache.sysds.runtime.instructions.cp.FunctionCallCPInstruction;
@@ -36,6 +33,13 @@ import org.apache.sysds.runtime.instructions.spark.SPInstruction;
 import org.apache.sysds.runtime.lineage.LineageCacheConfig.ReuseCacheType;
 import org.apache.sysds.runtime.lineage.LineageCacheStatistics;
 import org.apache.sysds.runtime.privacy.CheckedConstraintsLog;
+import org.apache.sysds.utils.stats.CodegenStatistics;
+import org.apache.sysds.utils.stats.HOPStatistics;
+import org.apache.sysds.utils.stats.NativeStatistics;
+import org.apache.sysds.utils.stats.ParamServStatistics;
+import org.apache.sysds.utils.stats.ParForStatistics;
+import org.apache.sysds.utils.stats.SparkStatistics;
+import org.apache.sysds.utils.stats.TransformStatistics;
 
 import java.lang.management.CompilationMXBean;
 import java.lang.management.GarbageCollectorMXBean;
@@ -86,125 +90,20 @@ public class Statistics
 	private static long jitCompileTime = 0; //in milli sec
 	private static long jvmGCTime = 0; //in milli sec
 	private static long jvmGCCount = 0; //count
-	
-	//HOP DAG recompile stats (potentially high update frequency)
-	private static final LongAdder hopRecompileTime = new LongAdder(); //in nano sec
-	private static final LongAdder hopRecompilePred = new LongAdder(); //count
-	private static final LongAdder hopRecompileSB = new LongAdder();   //count
 
-	//CODEGEN
-	private static final LongAdder codegenCompileTime = new LongAdder(); //in nano
-	private static final LongAdder codegenClassCompileTime = new LongAdder(); //in nano
-	private static final LongAdder codegenHopCompile = new LongAdder(); //count
-	private static final LongAdder codegenCPlanCompile = new LongAdder(); //count
-	private static final LongAdder codegenClassCompile = new LongAdder(); //count
-	private static final LongAdder codegenEnumAll = new LongAdder(); //count
-	private static final LongAdder codegenEnumAllP = new LongAdder(); //count
-	private static final LongAdder codegenEnumEval = new LongAdder(); //count
-	private static final LongAdder codegenEnumEvalP = new LongAdder(); //count
-	private static final LongAdder codegenOpCacheHits = new LongAdder(); //count
-	private static final LongAdder codegenOpCacheTotal = new LongAdder(); //count
-	private static final LongAdder codegenPlanCacheHits = new LongAdder(); //count
-	private static final LongAdder codegenPlanCacheTotal = new LongAdder(); //count
-	
 	//Function recompile stats 
 	private static final LongAdder funRecompileTime = new LongAdder(); //in nano sec
 	private static final LongAdder funRecompiles = new LongAdder(); //count
-	
-	//Spark-specific stats
-	private static long sparkCtxCreateTime = 0; 
-	private static final LongAdder sparkParallelize = new LongAdder();
-	private static final LongAdder sparkParallelizeCount = new LongAdder();
-	private static final LongAdder sparkCollect = new LongAdder();
-	private static final LongAdder sparkCollectCount = new LongAdder();
-	private static final LongAdder sparkBroadcast = new LongAdder();
-	private static final LongAdder sparkBroadcastCount = new LongAdder();
-	private static final LongAdder sparkAsyncPrefetchCount = new LongAdder();
-	private static final LongAdder sparkAsyncBroadcastCount = new LongAdder();
-
-	// Paramserv function stats (time is in milli sec)
-	private static final Timing psExecutionTimer = new Timing(false);
-	private static final LongAdder psExecutionTime = new LongAdder();
-	private static final LongAdder psNumWorkers = new LongAdder();
-	private static final LongAdder psSetupTime = new LongAdder();
-	private static final LongAdder psGradientComputeTime = new LongAdder();
-	private static final LongAdder psAggregationTime = new LongAdder();
-	private static final LongAdder psLocalModelUpdateTime = new LongAdder();
-	private static final LongAdder psModelBroadcastTime = new LongAdder();
-	private static final LongAdder psBatchIndexTime = new LongAdder();
-	private static final LongAdder psRpcRequestTime = new LongAdder();
-	private static final LongAdder psValidationTime = new LongAdder();
-	// Federated parameter server specifics (time is in milli sec)
-	private static final LongAdder fedPSDataPartitioningTime = new LongAdder();
-	private static final LongAdder fedPSWorkerComputingTime = new LongAdder();
-	private static final LongAdder fedPSGradientWeightingTime = new LongAdder();
-	private static final LongAdder fedPSCommunicationTime = new LongAdder();
-
-	//PARFOR optimization stats (low frequency updates)
-	private static long parforOptTime = 0; //in milli sec
-	private static long parforOptCount = 0; //count
-	private static long parforInitTime = 0; //in milli sec
-	private static long parforMergeTime = 0; //in milli sec
 
 	private static final LongAdder lTotalUIPVar = new LongAdder();
 	private static final LongAdder lTotalLix = new LongAdder();
 	private static final LongAdder lTotalLixUIP = new LongAdder();
 
-	// Transformencode stats
-	private static final LongAdder transformEncoderCount = new LongAdder();
-
-	//private static final LongAdder transformBuildTime = new LongAdder();
-	private static final LongAdder transformRecodeBuildTime = new LongAdder();
-	private static final LongAdder transformBinningBuildTime = new LongAdder();
-	private static final LongAdder transformImputeBuildTime = new LongAdder();
-
-	//private static final LongAdder transformApplyTime = new LongAdder();
-	private static final LongAdder transformRecodeApplyTime = new LongAdder();
-	private static final LongAdder transformDummyCodeApplyTime = new LongAdder();
-	private static final LongAdder transformPassThroughApplyTime = new LongAdder();
-	private static final LongAdder transformFeatureHashingApplyTime = new LongAdder();
-	private static final LongAdder transformBinningApplyTime = new LongAdder();
-	private static final LongAdder transformOmitApplyTime = new LongAdder();
-	private static final LongAdder transformImputeApplyTime = new LongAdder();
-
-
-	private static final LongAdder transformOutMatrixPreProcessingTime = new LongAdder();
-	private static final LongAdder transformOutMatrixPostProcessingTime = new LongAdder();
-
-	// Federated stats
-	private static final LongAdder federatedReadCount = new LongAdder();
-	private static final LongAdder federatedPutCount = new LongAdder();
-	private static final LongAdder federatedGetCount = new LongAdder();
-	private static final LongAdder federatedExecuteInstructionCount = new LongAdder();
-	private static final LongAdder federatedExecuteUDFCount = new LongAdder();
-	private static final LongAdder fedAsyncPrefetchCount = new LongAdder();
-
-	private static LongAdder numNativeFailures = new LongAdder();
-	public static LongAdder numNativeLibMatrixMultCalls = new LongAdder();
-	public static LongAdder numNativeConv2dCalls = new LongAdder();
-	public static LongAdder numNativeConv2dBwdDataCalls = new LongAdder();
-	public static LongAdder numNativeConv2dBwdFilterCalls = new LongAdder();
-	public static LongAdder numNativeSparseConv2dCalls = new LongAdder();
-	public static LongAdder numNativeSparseConv2dBwdFilterCalls = new LongAdder();
-	public static LongAdder numNativeSparseConv2dBwdDataCalls = new LongAdder();
-	public static long nativeLibMatrixMultTime = 0;
-	public static long nativeConv2dTime = 0;
-	public static long nativeConv2dBwdDataTime = 0;
-	public static long nativeConv2dBwdFilterTime = 0;
-	
 	public static long recomputeNNZTime = 0;
 	public static long examSparsityTime = 0;
 	public static long allocateDoubleArrTime = 0;
 
 	public static boolean allowWorkerStatistics = true;
-	
-	public static void incrementNativeFailuresCounter() {
-		numNativeFailures.increment();
-		// This is very rare and am not sure it is possible at all. Our initial experiments never encountered this case.
-		// Note: all the native calls have a fallback to Java; so if the user wants she can recompile SystemDS by
-		// commenting this exception and everything should work fine.
-		throw new RuntimeException("Unexpected ERROR: OOM caused during JNI transfer. Please disable native BLAS by setting enviroment variable: SYSTEMDS_BLAS=none");
-	}
 
 	public static long getNoOfExecutedSPInst() {
 		return numExecutedSPInst.longValue();
@@ -225,11 +124,7 @@ public class Statistics
 	public static void incrementNoOfCompiledSPInst() {
 		numCompiledSPInst.increment();
 	}
-	
-	public static boolean createdSparkContext() {
-		return sparkCtxCreateTime > 0;
-	}
-	
+
 	public static long getTotalUIPVar() {
 		return lTotalUIPVar.longValue();
 	}
@@ -279,124 +174,6 @@ public class Statistics
 	public static synchronized void incrementJVMgcCount( long delta ) {
 		jvmGCCount += delta;
 	}
-	
-	public static void incrementHOPRecompileTime( long delta ) {
-		hopRecompileTime.add(delta);
-	}
-	
-	public static void incrementHOPRecompilePred() {
-		hopRecompilePred.increment();
-	}
-	
-	public static void incrementHOPRecompilePred(long delta) {
-		hopRecompilePred.add(delta);
-	}
-	
-	public static void incrementHOPRecompileSB() {
-		hopRecompileSB.increment();
-	}
-	
-	public static void incrementHOPRecompileSB(long delta) {
-		hopRecompileSB.add(delta);
-	}
-	
-	public static void incrementCodegenDAGCompile() {
-		codegenHopCompile.increment();
-	}
-	
-	public static void incrementCodegenCPlanCompile(long delta) {
-		codegenCPlanCompile.add(delta);
-	}
-	
-	public static void incrementCodegenEnumAll(long delta) {
-		codegenEnumAll.add(delta);
-	}
-	public static void incrementCodegenEnumAllP(long delta) {
-		codegenEnumAllP.add(delta);
-	}
-	public static void incrementCodegenEnumEval(long delta) {
-		codegenEnumEval.add(delta);
-	}
-	public static void incrementCodegenEnumEvalP(long delta) {
-		codegenEnumEvalP.add(delta);
-	}
-	
-	public static void incrementCodegenClassCompile() {
-		codegenClassCompile.increment();
-	}
-	
-	public static void incrementCodegenCompileTime(long delta) {
-		codegenCompileTime.add(delta);
-	}
-	
-	public static void incrementCodegenClassCompileTime(long delta) {
-		codegenClassCompileTime.add(delta);
-	}
-	
-	public static void incrementCodegenOpCacheHits() {
-		codegenOpCacheHits.increment();
-	}
-	
-	public static void incrementCodegenOpCacheTotal() {
-		codegenOpCacheTotal.increment();
-	}
-	
-	public static void incrementCodegenPlanCacheHits() {
-		codegenPlanCacheHits.increment();
-	}
-	
-	public static void incrementCodegenPlanCacheTotal() {
-		codegenPlanCacheTotal.increment();
-	}
-	
-	public static long getCodegenDAGCompile() {
-		return codegenHopCompile.longValue();
-	}
-	
-	public static long getCodegenCPlanCompile() {
-		return codegenCPlanCompile.longValue();
-	}
-	
-	public static long getCodegenEnumAll() {
-		return codegenEnumAll.longValue();
-	}
-	public static long getCodegenEnumAllP() {
-		return codegenEnumAllP.longValue();
-	}
-	public static long getCodegenEnumEval() {
-		return codegenEnumEval.longValue();
-	}
-	public static long getCodegenEnumEvalP() {
-		return codegenEnumEvalP.longValue();
-	}
-	
-	public static long getCodegenClassCompile() {
-		return codegenClassCompile.longValue();
-	}
-	
-	public static long getCodegenCompileTime() {
-		return codegenCompileTime.longValue();
-	}
-	
-	public static long getCodegenClassCompileTime() {
-		return codegenClassCompileTime.longValue();
-	}
-	
-	public static long getCodegenOpCacheHits() {
-		return codegenOpCacheHits.longValue();
-	}
-	
-	public static long getCodegenOpCacheTotal() {
-		return codegenOpCacheTotal.longValue();
-	}
-	
-	public static long getCodegenPlanCacheHits() {
-		return codegenPlanCacheHits.longValue();
-	}
-	
-	public static long getCodegenPlanCacheTotal() {
-		return codegenPlanCacheTotal.longValue();
-	}
 
 	public static void incrementFunRecompileTime( long delta ) {
 		funRecompileTime.add(delta);
@@ -404,48 +181,6 @@ public class Statistics
 	
 	public static void incrementFunRecompiles() {
 		funRecompiles.increment();
-	}
-	
-	public static synchronized void incrementParForOptimCount(){
-		parforOptCount ++;
-	}
-	
-	public static synchronized void incrementParForOptimTime( long time ) {
-		parforOptTime += time;
-	}
-	
-	public static synchronized void incrementParForInitTime( long time ) {
-		parforInitTime += time;
-	}
-	
-	public static synchronized void incrementParForMergeTime( long time ) {
-		parforMergeTime += time;
-	}
-
-	public static synchronized void incFederated(RequestType rqt){
-		switch (rqt) {
-			case READ_VAR:
-				federatedReadCount.increment();
-				break;
-			case PUT_VAR:
-				federatedPutCount.increment();
-				break;
-			case GET_VAR:
-				federatedGetCount.increment();
-				break;
-			case EXEC_INST:
-				federatedExecuteInstructionCount.increment();
-				break;
-			case EXEC_UDF:
-				federatedExecuteUDFCount.increment();
-				break;
-			default:
-				break;
-		}
-	}
-
-	public static void incFedAsyncPrefetchCount(long c) {
-		fedAsyncPrefetchCount.add(c);
 	}
 
 	public static void startCompileTimer() {
@@ -489,42 +224,17 @@ public class Statistics
 	
 	public static void reset()
 	{
-		hopRecompileTime.reset();
-		hopRecompilePred.reset();
-		hopRecompileSB.reset();
-		
+		HOPStatistics.reset();
+
 		funRecompiles.reset();
 		funRecompileTime.reset();
-		
-		codegenHopCompile.reset();
-		codegenCPlanCompile.reset();
-		codegenClassCompile.reset();
-		codegenEnumAll.reset();
-		codegenEnumAllP.reset();
-		codegenEnumEval.reset();
-		codegenEnumEvalP.reset();
-		codegenCompileTime.reset();
-		codegenClassCompileTime.reset();
-		codegenOpCacheHits.reset();
-		codegenOpCacheTotal.reset();
-		codegenPlanCacheHits.reset();
-		codegenPlanCacheTotal.reset();
-		
-		parforOptCount = 0;
-		parforOptTime = 0;
-		parforInitTime = 0;
-		parforMergeTime = 0;
-		
-		sparkCtxCreateTime = 0;
-		sparkBroadcast.reset();
-		sparkBroadcastCount.reset();
-		sparkAsyncPrefetchCount.reset();
-		sparkAsyncBroadcastCount.reset();
-		sparkParallelize.reset();
-		sparkParallelizeCount.reset();
-		sparkCollect.reset();
-		sparkCollectCount.reset();
-		
+
+		CodegenStatistics.reset();
+
+		ParForStatistics.reset();
+
+		SparkStatistics.reset();
+
 		lTotalLix.reset();
 		lTotalLixUIP.reset();
 		lTotalUIPVar.reset();
@@ -538,24 +248,8 @@ public class Statistics
 		resetCPHeavyHitters();
 
 		GPUStatistics.reset();
-		numNativeLibMatrixMultCalls.reset();
-		numNativeSparseConv2dCalls.reset();
-		numNativeSparseConv2dBwdDataCalls.reset();
-		numNativeSparseConv2dBwdFilterCalls.reset();
-		numNativeConv2dCalls.reset();
-		numNativeConv2dBwdDataCalls.reset();
-		numNativeConv2dBwdFilterCalls.reset();
-		numNativeFailures.reset();
-		nativeLibMatrixMultTime = 0;
-		nativeConv2dTime = 0;
-		nativeConv2dBwdFilterTime = 0;
-		nativeConv2dBwdDataTime = 0;
-		federatedReadCount.reset();
-		federatedPutCount.reset();
-		federatedGetCount.reset();
-		federatedExecuteInstructionCount.reset();
-		federatedExecuteUDFCount.reset();
-		fedAsyncPrefetchCount.reset();
+
+		NativeStatistics.reset();
 
 		DMLCompressionStatistics.reset();
 	}
@@ -574,177 +268,6 @@ public class Statistics
 
 	public static void resetCPHeavyHitters(){
 		_instStats.clear();
-	}
-
-	public static void setSparkCtxCreateTime(long ns) {
-		sparkCtxCreateTime = ns;
-	}
-	
-	public static void accSparkParallelizeTime(long t) {
-		sparkParallelize.add(t);
-	}
-
-	public static void incSparkParallelizeCount(long c) {
-		sparkParallelizeCount.add(c);
-	}
-
-	public static void accSparkCollectTime(long t) {
-		sparkCollect.add(t);
-		sparkCollectCount.add(1);
-	}
-
-	public static long getSparkCollectCount(){
-		return sparkCollectCount.longValue();
-	}
-
-	public static void accSparkBroadCastTime(long t) {
-		sparkBroadcast.add(t);
-	}
-
-	public static void incSparkBroadcastCount(long c) {
-		sparkBroadcastCount.add(c);
-	}
-
-	public static void incSparkAsyncPrefetchCount(long c) {
-		sparkAsyncPrefetchCount.add(c);
-	}
-
-	public static void incSparkAsyncBroadcastCount(long c) {
-		sparkAsyncBroadcastCount.add(c);
-	}
-
-	public static void incWorkerNumber() {
-		psNumWorkers.increment();
-	}
-
-	public static void incWorkerNumber(long n) {
-		psNumWorkers.add(n);
-	}
-
-	public static Timing getPSExecutionTimer() {
-		return psExecutionTimer;
-	}
-
-	public static double getPSExecutionTime() {
-		return psExecutionTime.doubleValue();
-	}
-
-	public static void accPSExecutionTime(long n) {
-		psExecutionTime.add(n);
-	}
-
-	public static void accPSSetupTime(long t) {
-		psSetupTime.add(t);
-	}
-
-	public static void accPSGradientComputeTime(long t) {
-		psGradientComputeTime.add(t);
-	}
-
-	public static void accPSAggregationTime(long t) {
-		psAggregationTime.add(t);
-	}
-
-	public static void accPSLocalModelUpdateTime(long t) {
-		psLocalModelUpdateTime.add(t);
-	}
-
-	public static void accPSModelBroadcastTime(long t) {
-		psModelBroadcastTime.add(t);
-	}
-
-	public static void accPSBatchIndexingTime(long t) {
-		psBatchIndexTime.add(t);
-	}
-
-	public static void accPSRpcRequestTime(long t) {
-		psRpcRequestTime.add(t);
-	}
-
-	public static double getPSValidationTime() {
-		return psValidationTime.doubleValue();
-	}
-
-	public static void accPSValidationTime(long t) {
-		psValidationTime.add(t);
-	}
-
-	public static void accFedPSDataPartitioningTime(long t) {
-		fedPSDataPartitioningTime.add(t);
-	}
-
-	public static void accFedPSWorkerComputing(long t) {
-		fedPSWorkerComputingTime.add(t);
-	}
-
-	public static void accFedPSGradientWeightingTime(long t) {
-		fedPSGradientWeightingTime.add(t);
-	}
-
-	public static void accFedPSCommunicationTime(long t) { fedPSCommunicationTime.add(t);}
-
-	public static void incTransformEncoderCount(long encoders){
-		transformEncoderCount.add(encoders);
-	}
-
-	public static void incTransformRecodeApplyTime(long t){
-		transformRecodeApplyTime.add(t);
-	}
-
-	public static void incTransformDummyCodeApplyTime(long t){
-		transformDummyCodeApplyTime.add(t);
-	}
-
-	public static void incTransformBinningApplyTime(long t){
-		transformBinningApplyTime.add(t);
-	}
-
-	public static void incTransformPassThroughApplyTime(long t){
-		transformPassThroughApplyTime.add(t);
-	}
-
-	public static void incTransformFeatureHashingApplyTime(long t){
-		transformFeatureHashingApplyTime.add(t);
-	}
-
-	public static void incTransformOmitApplyTime(long t) {
-		transformOmitApplyTime.add(t);
-	}
-
-	public static void incTransformImputeApplyTime(long t) {
-		transformImputeApplyTime.add(t);
-	}
-
-	public static void incTransformRecodeBuildTime(long t){
-		transformRecodeBuildTime.add(t);
-	}
-
-	public static void incTransformBinningBuildTime(long t){
-		transformBinningBuildTime.add(t);
-	}
-
-	public static void incTransformImputeBuildTime(long t) {
-		transformImputeBuildTime.add(t);
-	}
-
-	public static void incTransformOutMatrixPreProcessingTime(long t){
-		transformOutMatrixPreProcessingTime.add(t);
-	}
-
-	public static void incTransformOutMatrixPostProcessingTime(long t){
-		transformOutMatrixPostProcessingTime.add(t);
-	}
-
-	public static long getTransformEncodeBuildTime(){
-		return transformBinningBuildTime.longValue() + transformImputeBuildTime.longValue() +
-				transformRecodeBuildTime.longValue();
-	}
-
-	public static long getTransformEncodeApplyTime(){
-		return transformDummyCodeApplyTime.longValue() + transformBinningApplyTime.longValue() +
-				transformFeatureHashingApplyTime.longValue() + transformPassThroughApplyTime.longValue() +
-				transformRecodeApplyTime.longValue() + transformOmitApplyTime.longValue() +
-				transformImputeApplyTime.longValue();
 	}
 
 	public static String getCPHeavyHitterCode( Instruction inst )
@@ -1031,19 +554,7 @@ public class Statistics
 		
 		return ret;
 	}
-	
-	public static long getHopRecompileTime(){
-		return hopRecompileTime.longValue();
-	}
-	
-	public static long getHopRecompiledPredDAGs(){
-		return hopRecompilePred.longValue();
-	}
-	
-	public static long getHopRecompiledSBDAGs(){
-		return hopRecompileSB.longValue();
-	}
-	
+
 	public static long getFunRecompileTime(){
 		return funRecompileTime.longValue();
 	}
@@ -1051,34 +562,10 @@ public class Statistics
 	public static long getFunRecompiles(){
 		return funRecompiles.longValue();
 	}
-		
-	public static long getParforOptCount(){
-		return parforOptCount;
-	}
-	
-	public static long getParforOptTime(){
-		return parforOptTime;
-	}
-	
-	public static long getParforInitTime(){
-		return parforInitTime;
-	}
-	
-	public static long getParforMergeTime(){
-		return parforMergeTime;
-	}
 
 	public static long getNumPinnedObjects() { return maxNumPinnedObjects; }
 
 	public static double getSizeofPinnedObjects() { return maxSizeofPinnedObjects; }
-	
-	public static long getAsyncPrefetchCount() {
-		return sparkAsyncPrefetchCount.longValue();
-	}
-
-	public static long getAsyncBroadcastCount() {
-		return sparkAsyncBroadcastCount.longValue();
-	}
 
 	/**
 	 * Returns statistics of the DML program that was recently completed as a string
@@ -1126,18 +613,9 @@ public class Statistics
 		//show extended caching/compilation statistics
 		if( DMLScript.STATISTICS )
 		{
-			if(NativeHelper.CURRENT_NATIVE_BLAS_STATE == NativeHelper.NativeBlasState.SUCCESSFULLY_LOADED_NATIVE_BLAS_AND_IN_USE) {
-				String blas = NativeHelper.getCurrentBLAS();
-				sb.append("Native " + blas + " calls (dense mult/conv/bwdF/bwdD):\t" + numNativeLibMatrixMultCalls.longValue()  + "/" +
-						numNativeConv2dCalls.longValue() + "/" + numNativeConv2dBwdFilterCalls.longValue()
-						+ "/" + numNativeConv2dBwdDataCalls.longValue() + ".\n");
-				sb.append("Native " + blas + " calls (sparse conv/bwdF/bwdD):\t" +
-						numNativeSparseConv2dCalls.longValue() + "/" + numNativeSparseConv2dBwdFilterCalls.longValue()
-						+ "/" + numNativeSparseConv2dBwdDataCalls.longValue() + ".\n");
-				sb.append("Native " + blas + " times (dense mult/conv/bwdF/bwdD):\t" + String.format("%.3f", nativeLibMatrixMultTime*1e-9) + "/" +
-						String.format("%.3f", nativeConv2dTime*1e-9) + "/" + String.format("%.3f", nativeConv2dBwdFilterTime*1e-9) + "/" +
-						String.format("%.3f", nativeConv2dBwdDataTime*1e-9) + ".\n");
-			}
+			if(NativeHelper.CURRENT_NATIVE_BLAS_STATE == NativeHelper.NativeBlasState.SUCCESSFULLY_LOADED_NATIVE_BLAS_AND_IN_USE)
+				sb.append(NativeStatistics.displayNativeStatistics());
+
 			if(recomputeNNZTime != 0 || examSparsityTime != 0 || allocateDoubleArrTime != 0) {
 				sb.append("MatrixBlock times (recomputeNNZ/examSparsity/allocateDoubleArr):\t" + String.format("%.3f", recomputeNNZTime*1e-9) + "/" +
 						String.format("%.3f", examSparsityTime*1e-9) + "/" + String.format("%.3f", allocateDoubleArrTime*1e-9)  + ".\n");
@@ -1148,8 +626,9 @@ public class Statistics
 			sb.append("Cache times (ACQr/m, RLS, EXP):\t" + CacheStatistics.displayTime() + " sec.\n");
 			if (DMLScript.JMLC_MEM_STATISTICS)
 				sb.append("Max size of live objects:\t" + byteCountToDisplaySize(getSizeofPinnedObjects()) + " ("  + getNumPinnedObjects() + " total objects)" + "\n");
-			sb.append("HOP DAGs recompiled (PRED, SB):\t" + getHopRecompiledPredDAGs() + "/" + getHopRecompiledSBDAGs() + ".\n");
-			sb.append("HOP DAGs recompile time:\t" + String.format("%.3f", ((double)getHopRecompileTime())/1000000000) + " sec.\n");
+
+			sb.append(HOPStatistics.displayHOPStatistics());
+
 			if( getFunRecompiles()>0 ) {
 				sb.append("Functions recompiled:\t\t" + getFunRecompiles() + ".\n");
 				sb.append("Functions recompile time:\t" + String.format("%.3f", ((double)getFunRecompileTime())/1000000000) + " sec.\n");
@@ -1163,116 +642,21 @@ public class Statistics
 				sb.append("LinCache Computetime (S/M): \t" + LineageCacheStatistics.displayComputeTime() + " sec.\n");
 				sb.append("LinCache Rewrites:    \t\t" + LineageCacheStatistics.displayRewrites() + ".\n");
 			}
-			if( ConfigurationManager.isCodegenEnabled() ) {
-				sb.append("Codegen compile (DAG,CP,JC):\t" + getCodegenDAGCompile() + "/"
-						+ getCodegenCPlanCompile() + "/" + getCodegenClassCompile() + ".\n");
-				sb.append("Codegen enum (ALLt/p,EVALt/p):\t" + getCodegenEnumAll() + "/" +
-						getCodegenEnumAllP() + "/" + getCodegenEnumEval() + "/" + getCodegenEnumEvalP() + ".\n");
-				sb.append("Codegen compile times (DAG,JC):\t" + String.format("%.3f", (double)getCodegenCompileTime()/1000000000) + "/" +
-						String.format("%.3f", (double)getCodegenClassCompileTime()/1000000000)  + " sec.\n");
-				sb.append("Codegen enum plan cache hits:\t" + getCodegenPlanCacheHits() + "/" + getCodegenPlanCacheTotal() + ".\n");
-				sb.append("Codegen op plan cache hits:\t" + getCodegenOpCacheHits() + "/" + getCodegenOpCacheTotal() + ".\n");
-			}
-			if( OptimizerUtils.isSparkExecutionMode() ){
-				String lazy = SparkExecutionContext.isLazySparkContextCreation() ? "(lazy)" : "(eager)";
-				sb.append("Spark ctx create time "+lazy+":\t"+
-						String.format("%.3f", sparkCtxCreateTime*1e-9)  + " sec.\n" ); // nanoSec --> sec
-				sb.append("Spark trans counts (par,bc,col):" +
-						String.format("%d/%d/%d.\n", sparkParallelizeCount.longValue(),
-								sparkBroadcastCount.longValue(), sparkCollectCount.longValue()));
-				sb.append("Spark trans times (par,bc,col):\t" +
-						String.format("%.3f/%.3f/%.3f secs.\n",
-								sparkParallelize.longValue()*1e-9,
-								sparkBroadcast.longValue()*1e-9,
-								sparkCollect.longValue()*1e-9));
-				if (OptimizerUtils.ASYNC_TRIGGER_RDD_OPERATIONS)
-					sb.append("Spark async. count (pf,bc): \t" + 
-							String.format("%d/%d.\n", getAsyncPrefetchCount(), getAsyncBroadcastCount()));
-			}
-			if (psNumWorkers.longValue() > 0) {
-				sb.append(String.format("Paramserv total execution time:\t%.3f secs.\n", psExecutionTime.doubleValue() / 1000));
-				sb.append(String.format("Paramserv total num workers:\t%d.\n", psNumWorkers.longValue()));
-				sb.append(String.format("Paramserv setup time:\t\t%.3f secs.\n", psSetupTime.doubleValue() / 1000));
 
-				if(fedPSDataPartitioningTime.doubleValue() > 0) { 	//if data partitioning happens this is the federated case
-					sb.append(String.format("PS fed data partitioning time:\t%.3f secs.\n", fedPSDataPartitioningTime.doubleValue() / 1000));
-					sb.append(String.format("PS fed comm time (cum):\t\t%.3f secs.\n", fedPSCommunicationTime.doubleValue() / 1000));
-					sb.append(String.format("PS fed worker comp time (cum):\t%.3f secs.\n", fedPSWorkerComputingTime.doubleValue() / 1000));
-					sb.append(String.format("PS fed grad. weigh. time (cum):\t%.3f secs.\n", fedPSGradientWeightingTime.doubleValue() / 1000));
-					sb.append(String.format("PS fed global model agg time:\t%.3f secs.\n", psAggregationTime.doubleValue() / 1000));
-				}
-				else {
-					sb.append(String.format("Paramserv grad compute time:\t%.3f secs.\n", psGradientComputeTime.doubleValue() / 1000));
-					sb.append(String.format("Paramserv model update time:\t%.3f/%.3f secs.\n",
-						psLocalModelUpdateTime.doubleValue() / 1000, psAggregationTime.doubleValue() / 1000));
-					sb.append(String.format("Paramserv model broadcast time:\t%.3f secs.\n", psModelBroadcastTime.doubleValue() / 1000));
-					sb.append(String.format("Paramserv batch slice time:\t%.3f secs.\n", psBatchIndexTime.doubleValue() / 1000));
-					sb.append(String.format("Paramserv RPC request time:\t%.3f secs.\n", psRpcRequestTime.doubleValue() / 1000));
-				}
-				sb.append(String.format("Paramserv valdiation time:\t%.3f secs.\n", psValidationTime.doubleValue() / 1000));
-			}
-			if( parforOptCount>0 ){
-				sb.append("ParFor loops optimized:\t\t" + getParforOptCount() + ".\n");
-				sb.append("ParFor optimize time:\t\t" + String.format("%.3f", ((double)getParforOptTime())/1000) + " sec.\n");
-				sb.append("ParFor initialize time:\t\t" + String.format("%.3f", ((double)getParforInitTime())/1000) + " sec.\n");
-				sb.append("ParFor result merge time:\t" + String.format("%.3f", ((double)getParforMergeTime())/1000) + " sec.\n");
-				sb.append("ParFor total update in-place:\t" + lTotalUIPVar + "/" + lTotalLixUIP + "/" + lTotalLix + "\n");
-			}
-			if( federatedReadCount.longValue() > 0){
-				sb.append("Federated I/O (Read, Put, Get):\t" + 
-					federatedReadCount.longValue() + "/" +
-					federatedPutCount.longValue() + "/" +
-					federatedGetCount.longValue() + ".\n");
-				sb.append("Federated Execute (Inst, UDF):\t" +
-					federatedExecuteInstructionCount.longValue() + "/" +
-					federatedExecuteUDFCount.longValue() + ".\n");
-				sb.append("Federated prefetch count:\t" +
-					fedAsyncPrefetchCount.longValue() + ".\n");
-			}
-			if( transformEncoderCount.longValue() > 0) {
-				//TODO: Cleanup and condense
-				sb.append("TransformEncode num. encoders:\t").append(transformEncoderCount.longValue()).append("\n");
-				sb.append("TransformEncode build time:\t").append(String.format("%.3f",
-						getTransformEncodeBuildTime()*1e-9)).append(" sec.\n");
-				if(transformRecodeBuildTime.longValue() > 0)
-					sb.append("\tRecode build time:\t").append(String.format("%.3f",
-							transformRecodeBuildTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformBinningBuildTime.longValue() > 0)
-					sb.append("\tBinning build time:\t").append(String.format("%.3f",
-							transformBinningBuildTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformImputeBuildTime.longValue() > 0)
-					sb.append("\tImpute build time:\t").append(String.format("%.3f",
-							transformImputeBuildTime.longValue()*1e-9)).append(" sec.\n");
+			if( ConfigurationManager.isCodegenEnabled() )
+				sb.append(CodegenStatistics.displayCodegenStatistics());
 
-				sb.append("TransformEncode apply time:\t").append(String.format("%.3f",
-						getTransformEncodeApplyTime()*1e-9)).append(" sec.\n");
-				if(transformRecodeApplyTime.longValue() > 0)
-					sb.append("\tRecode apply time:\t").append(String.format("%.3f",
-							transformRecodeApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformBinningApplyTime.longValue() > 0)
-					sb.append("\tBinning apply time:\t").append(String.format("%.3f",
-							transformBinningApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformDummyCodeApplyTime.longValue() > 0)
-					sb.append("\tDummyCode apply time:\t").append(String.format("%.3f",
-							transformDummyCodeApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformFeatureHashingApplyTime.longValue() > 0)
-					sb.append("\tHashing apply time:\t").append(String.format("%.3f",
-							transformFeatureHashingApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformPassThroughApplyTime.longValue() > 0)
-					sb.append("\tPassThrough apply time:\t").append(String.format("%.3f",
-							transformPassThroughApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformOmitApplyTime.longValue() > 0)
-					sb.append("\tOmit apply time:\t").append(String.format("%.3f",
-							transformOmitApplyTime.longValue()*1e-9)).append(" sec.\n");
-				if(transformImputeApplyTime.longValue() > 0)
-					sb.append("\tImpute apply time:\t").append(String.format("%.3f",
-							transformImputeApplyTime.longValue()*1e-9)).append(" sec.\n");
+			if( OptimizerUtils.isSparkExecutionMode() )
+				sb.append(SparkStatistics.displaySparkStatistics());
 
-				sb.append("TransformEncode PreProc. time:\t").append(String.format("%.3f",
-						transformOutMatrixPreProcessingTime.longValue()*1e-9)).append(" sec.\n");
-				sb.append("TransformEncode PostProc. time:\t").append(String.format("%.3f",
-						transformOutMatrixPostProcessingTime.longValue()*1e-9)).append(" sec.\n");
-			}
+			sb.append(ParamServStatistics.displayParamServStatistics());
+
+			sb.append(ParForStatistics.displayParForStatistics());
+
+			sb.append(FederatedStatistics.displayFedIOExecStatistics());
+			sb.append(FederatedStatistics.displayFedLookupTableStats());
+
+			sb.append(TransformStatistics.displayTransformStatistics());
 
 			if(ConfigurationManager.isCompressionEnabled()){
 				DMLCompressionStatistics.display(sb);
